+++ src/config.py
@@ -5,4 +5,4 @@ This module contains constants used across the application.
 """
 
 APP_NAME = "Skunkworks"
-APP_VERSION = "1.0.2"
+APP_VERSION = "1.0.3"